# spring-server/src/main/java/ApiNegocio/model/GetRequest.java
package ApiNegocio.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import org.springframework.validation.annotation.Validated;
import javax.validation.Valid;
import javax.validation.constraints.*;

/**
 * GetRequest
 */
@Validated
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-07-31T15:25:59.597Z")

public class GetRequest   {
  @JsonProperty("id")
  private String id = null;

  @JsonProperty("parametro")
  private String parametro = null;

  @JsonProperty("token")
  private String token = null;

  public GetRequest id(String id) {
    this.id = id;
    return this;
  }

  /**
   * Get tipoConsulta
   * @return tipoConsulta
  **/
  @ApiModelProperty(required = true, value = "")
  @NotNull


  public String getID() {
    return id;
  }

  public void setID(String id) {
    this.id = id;
  }

  public GetRequest parametro(String parametro) {
    this.parametro = parametro;
    return this;
  }

  /**
   * Get parametro
   * @return parametro
  **/
  @ApiModelProperty(required = true, value = "")
  @NotNull


  public String getParametro() {
    return parametro;
  }

  public void setParametro(String parametro) {
    this.parametro = parametro;
  }

  public GetRequest token(String token) {
    this.token = token;
    return this;
  }

  /**
   * Get token
   * @return token
  **/
  @ApiModelProperty(required = true, value = "")
  @NotNull


  public String getToken() {
    return token;
  }

  public void setToken(String token) {
    this.token = token;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    GetRequest getRequest = (GetRequest) o;
    return Objects.equals(this.id, getRequest.id) &&
        Objects.equals(this.parametro, getRequest.parametro) &&
        Objects.equals(this.token, getRequest.token);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, parametro, token);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class GetRequest {\n");
    
    sb.append("    id: ").append(toIndentedString(id)).append("\n");
    sb.append("    parametro: ").append(toIndentedString(parametro)).append("\n");
    sb.append("    token: ").append(toIndentedString(token)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}